Toggle favorite on restaurant index
